Add settings dropdown

+++ frontend/pages/account/account-details.vue
@@ -2,7 +2,9 @@
   <div>
     <breadcrumbs :routes="breadcrumbs" />
     <div class="account container tablet-width padding-horizontal--large">
-      <h1 class="margin-top--xxlarge margin-bottom--large">My account details</h1>
+      <h1 class="margin-top--xxlarge margin-bottom--large">
+        My account details
+      </h1>
       <nuxt-link to="/account/public-profile-view/">
         See your public profile
         <right-arrow class="icon icon--blue" />
@@ -57,24 +59,42 @@
 
         <h3>Recipes connected to your account</h3>
         <dl class="flex-row">
-          <account-link-list title="My recipes:" :links="userRecipes" basePath="/recipes/" />
+          <account-link-list
+            title="My recipes:"
+            :links="userRecipes"
+            basePath="/recipes/"
+          />
           <account-link-list
             title="Recipes shared with me:"
             :links="sharedRecipes"
             basePath="/recipes/"
           />
-          <account-link-list title="My recipe links:" :links="recipeLinks" :externalURL="true" />
+          <account-link-list
+            title="My recipe links:"
+            :links="recipeLinks"
+            :externalURL="true"
+          />
         </dl>
 
         <h3>Cooks connected to your account</h3>
         <dl class="flex-row">
-          <account-link-list title="Following:" :links="cooksFollowed" basePath="/cooks/" />
-          <account-link-list title="Followers:" :links="cookFollowers" basePath="/cooks/" />
+          <account-link-list
+            title="Following:"
+            :links="cooksFollowed"
+            basePath="/cooks/"
+          />
+          <account-link-list
+            title="Followers:"
+            :links="cookFollowers"
+            basePath="/cooks/"
+          />
         </dl>
         <button
           class="button button--small button--transparent button--transparent-red margin-top--large"
           @click="toggleAlert"
-        >Delete my account</button>
+        >
+          Delete my account
+        </button>
         <p class="system-message">{{ systemMessage }}</p>
       </div>
     </div>
@@ -255,23 +275,27 @@ export default {
       let fileName = this.photoURL;
 
       var storageRef = this.$fireStorage.ref();
-      var profileImgRef = storageRef.child(`images/users/${userID}/profileImg`);
+      if (userID) {
+        var profileImgRef = storageRef.child(
+          `images/users/${userID}/profileImg`
+        );
 
-      profileImgRef
-        .listAll()
-        .then(function(res) {
-          res.items.forEach(function(itemRef) {
-            itemRef.delete();
+        profileImgRef
+          .listAll()
+          .then(function(res) {
+            res.items.forEach(function(itemRef) {
+              itemRef.delete();
+            });
+          })
+          .then(() => {
+            componentThis.removeProfileImageFromDb();
+          })
+          .catch(function(error) {
+            console.log("Error deleting files:", error.message);
+            componentThis.removeProfileImgSystemMessage =
+              "We're having trouble deleting your profile image. Please try again later or contact us.";
           });
-        })
-        .then(() => {
-          componentThis.removeProfileImageFromDb();
-        })
-        .catch(function(error) {
-          console.log("Error deleting files:", error.message);
-          componentThis.removeProfileImgSystemMessage =
-            "We're having trouble deleting your profile image. Please try again later or contact us.";
-        });
+      }
     },
     removeProfileImageFromDb() {
       let componentThis = this;
